refactor: reorganize media component imports and move to media subdir for domain organisation
